# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "vLLM Docker API",
    "description": "A TypeScript Node.js API for managing vLLM containers with OpenAI-compatible endpoints and comprehensive model management capabilities.",
    "version": "1.0.0",
    "contact": {
      "name": "vLLM Docker API",
      "url": "https://github.com/your-repo/vllm-docker-api"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/start": {
      "post": {
        "summary": "Start vLLM container",
        "description": "Starts a vLLM container with the specified model or default model.",
        "operationId": "startContainer",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              },
              "example": {
                "model": "mistralai/Mistral-7B-Instruct-v0.3"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Container started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stop": {
      "post": {
        "summary": "Stop vLLM container",
        "description": "Stops the running vLLM container.",
        "operationId": "stopContainer",
        "responses": {
          "200": {
            "description": "Container stopped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/remove": {
      "delete": {
        "summary": "Remove vLLM container",
        "description": "Stops and removes the vLLM container.",
        "operationId": "removeContainer",
        "responses": {
          "200": {
            "description": "Container removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/status": {
      "get": {
        "summary": "Get container status",
        "description": "Returns the current status of the vLLM container.",
        "operationId": "getContainerStatus",
        "responses": {
          "200": {
            "description": "Status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/download": {
      "post": {
        "summary": "Download model from Hugging Face",
        "description": "Downloads a model from Hugging Face to the local models directory.",
        "operationId": "downloadModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadModelRequest"
              },
              "example": {
                "model": "mistralai/Mistral-7B-Instruct-v0.3"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Download started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - model already exists or invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Download failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/download/progress": {
      "get": {
        "summary": "Get download progress",
        "description": "Returns download progress for all models or a specific model.",
        "operationId": "getDownloadProgress",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "description": "Specific model ID to check progress for. If not provided, returns progress for all downloads.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "mistralai/Mistral-7B-Instruct-v0.3"
          }
        ],
        "responses": {
          "200": {
            "description": "Progress retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DownloadProgressResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AllDownloadsProgressResponse"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Model download not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/download/progress/{model}": {
      "get": {
        "summary": "Get download progress for specific model",
        "description": "Returns download progress for a specific model.",
        "operationId": "getModelDownloadProgress",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "description": "Model ID to check progress for",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "mistralai/Mistral-7B-Instruct-v0.3"
          }
        ],
        "responses": {
          "200": {
            "description": "Progress retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadProgressResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model download not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/download/{model}": {
      "delete": {
        "summary": "Cancel model download",
        "description": "Cancels an ongoing model download.",
        "operationId": "cancelModelDownload",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "description": "Model ID to cancel download for",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "mistralai/Mistral-7B-Instruct-v0.3"
          }
        ],
        "responses": {
          "200": {
            "description": "Download cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelDownloadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Download cannot be cancelled (already completed/failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Download not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "StartRequest": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "description": "Model to load. If not provided, uses the default model"
          }
        }
      },
      "StartResponse": {
        "type": "object",
        "required": ["ok", "model"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "model": {
            "type": "string",
            "description": "The model that was loaded"
          }
        }
      },
      "StopResponse": {
        "type": "object",
        "required": ["ok", "stopped"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "stopped": {
            "type": "boolean",
            "description": "Whether the container was stopped"
          },
          "message": {
            "type": "string",
            "description": "Additional message if not found or already stopped"
          }
        }
      },
      "RemoveResponse": {
        "type": "object",
        "required": ["ok", "removed"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "removed": {
            "type": "boolean",
            "description": "Whether the container was removed"
          },
          "message": {
            "type": "string",
            "description": "Additional message if not found"
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "required": ["container", "running", "port", "image"],
        "properties": {
          "container": {
            "type": "string",
            "description": "Container name"
          },
          "running": {
            "type": "boolean",
            "description": "Whether the container is running"
          },
          "port": {
            "type": "integer",
            "description": "Container port"
          },
          "image": {
            "type": "string",
            "description": "Docker image name"
          }
        }
      },
...
      "DownloadModelRequest": {
        "type": "object",
        "required": ["model"],
        "properties": {
          "model": {
            "type": "string",
            "description": "Hugging Face model ID (e.g., 'mistralai/Mistral-7B-Instruct-v0.3')"
          }
        }
      },
      "DownloadResponse": {
        "type": "object",
        "required": ["success", "message", "model"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Path where the model was downloaded"
          }
        }
      },
      "DownloadProgressResponse": {
        "type": "object",
        "required": ["success", "progress"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "progress": {
            "type": "object",
            "properties": {
              "model": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": ["downloading", "completed", "failed", "cancelled"]
              },
              "progress": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100,
                "description": "Download progress percentage"
              },
              "message": {
                "type": "string"
              },
              "startTime": {
                "type": "integer",
                "description": "Unix timestamp when download started"
              },
              "endTime": {
                "type": "integer",
                "description": "Unix timestamp when download ended (if completed)"
              },
              "path": {
                "type": "string",
                "description": "Path where the model is being downloaded"
              },
              "error": {
                "type": "string",
                "description": "Error message if download failed"
              },
              "duration": {
                "type": "integer",
                "description": "Download duration in milliseconds"
              }
            }
          }
        }
      },
      "AllDownloadsProgressResponse": {
        "type": "object",
        "required": ["success", "downloads"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "downloads": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "model": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": ["downloading", "completed", "failed", "cancelled"]
                },
                "progress": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 100,
                  "description": "Download progress percentage"
                },
                "message": {
                  "type": "string"
                },
                "startTime": {
                  "type": "integer",
                  "description": "Unix timestamp when download started"
                },
                "endTime": {
                  "type": "integer",
                  "description": "Unix timestamp when download ended (if completed)"
                },
                "path": {
                  "type": "string",
                  "description": "Path where the model is being downloaded"
                },
                "error": {
                  "type": "string",
                  "description": "Error message if download failed"
                },
                "duration": {
                  "type": "integer",
                  "description": "Download duration in milliseconds"
                }
              }
            }
          }
        }
      },
      "CancelDownloadResponse": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "description": "Zod validation errors"
          }
        }
      },
...
    }
  },
  "security": [],
...
}